Add class BadRequest

+++ VillaApplication/Controller/OwnerController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using VillaApplication.Model.Bo;
+using VillaApplication.Model.Data;
 using VillaApplication.Model.Dto;
 using VillaApplication.Service;
 
@@ -13,6 +14,7 @@ namespace OwnerApplication.Controller
         private readonly ILogger logger = logger;
 
         [HttpGet]
+        [Route("all")]
         public ActionResult<IEnumerable<OwnerDTO>> GetOwners()
         {
             return Ok(service.GetAllEntities());
@@ -21,11 +23,18 @@ namespace OwnerApplication.Controller
         [HttpGet("{id:int}")]
         public ActionResult<OwnerDTO> GetOwners(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest(new ErrorResponse(400, $"Id: {id} not valid."));
+            }
+
             OwnerDTO? villaDTO = service.GetById(id);
 
             if (villaDTO == null)
             {
-                return NotFound();
+                logger.LogError("Error to get Owner entity. Owner with id {Id} not found.", id);
+
+                return NotFound(new ErrorResponse(404, $"Owner with id: {id} not found."));
             }
             else
             {
@@ -36,21 +45,17 @@ namespace OwnerApplication.Controller
         [HttpPost]
         public ActionResult<OwnerDTO> CreateOwner([FromBody] OwnerBO bo)
         {
-            if (!ModelState.IsValid)
-            {
-                var messages = ModelState
-                  .SelectMany(modelState => modelState.Value.Errors)
-                  .Select(err => err.ErrorMessage)
-                  .ToList();
-
-                return BadRequest(messages);
-            }
             return Ok(service.AddOwner(bo));
         }
 
         [HttpDelete("{id:int}")]
         public ActionResult DeleteOwner(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest(new ErrorResponse(400, $"Id: {id} not valid."));
+            }
+
             bool deleted = service.Delete(id);
 
             if (deleted)
@@ -59,7 +64,7 @@ namespace OwnerApplication.Controller
             }
             else
             {
-                return BadRequest();
+                return NotFound(new ErrorResponse(404, $"Owner with id: {id} not found."));
             }
         }
 
@@ -69,6 +74,7 @@ namespace OwnerApplication.Controller
             if (id == 0 || bo == null)
             {
                 logger.LogError("Error to edit Vila.");
+
                 return BadRequest();
             }
